requirements.txt
-i https://pypi.org/simple
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.2
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3 ; python_version >= '3.7'
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
mako==1.2.4
markupsafe==2.1.3
packaging==23.1
passlib[bcrypt]==1.7.4
pluggy==1.2.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.3
pydantic-core==2.3.0
pytest==7.4.0
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlalchemy==2.0.19
sqlalchemy-utils==0.41.1
starlette==0.27.0
tomli==2.0.1
typer==0.9.0
typing-extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.1
